# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalystStreamRequest": {
        "properties": {
          "atr_floor": {
            "default": 0.1,
            "maximum": 0.3,
            "minimum": 0.0,
            "title": "Atr Floor",
            "type": "number"
          },
          "atr_multiplier": {
            "default": 3.0,
            "maximum": 6.0,
            "minimum": 0.5,
            "title": "Atr Multiplier",
            "type": "number"
          },
          "atr_period": {
            "default": 14,
            "maximum": 60.0,
            "minimum": 4.0,
            "title": "Atr Period",
            "type": "integer"
          },
          "commitment_curve": {
            "default": "linear",
            "enum": [
              "linear",
              "sqrt",
              "off"
            ],
            "title": "Commitment Curve",
            "type": "string"
          },
          "force_refresh": {
            "default": false,
            "title": "Force Refresh",
            "type": "boolean"
          },
          "k_sigma": {
            "default": 0.5,
            "maximum": 1.5,
            "minimum": 0.1,
            "title": "K Sigma",
            "type": "number"
          },
          "log_tol_fib": {
            "default": 0.05,
            "maximum": 0.3,
            "minimum": 0.01,
            "title": "Log Tol Fib",
            "type": "number"
          },
          "min_bars_between": {
            "default": 4,
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Min Bars Between",
            "type": "integer"
          },
          "mode": {
            "default": "explanation",
            "enum": [
              "explanation",
              "outlook",
              "risk",
              "differentiator"
            ],
            "title": "Mode",
            "type": "string"
          },
          "period": {
            "default": "max",
            "enum": [
              "2y",
              "5y",
              "10y",
              "max"
            ],
            "title": "Period",
            "type": "string"
          },
          "pivot_window": {
            "default": 2,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Pivot Window",
            "type": "integer"
          },
          "pull_depth_hi": {
            "default": 0.618,
            "maximum": 0.95,
            "minimum": 0.0,
            "title": "Pull Depth Hi",
            "type": "number"
          },
          "pull_depth_lo": {
            "default": 0.382,
            "maximum": 0.95,
            "minimum": 0.0,
            "title": "Pull Depth Lo",
            "type": "number"
          },
          "pull_depth_tol": {
            "default": 0.15,
            "maximum": 0.5,
            "minimum": 0.01,
            "title": "Pull Depth Tol",
            "type": "number"
          },
          "rate_tps": {
            "default": 40.0,
            "exclusiveMinimum": 0.0,
            "maximum": 500.0,
            "title": "Rate Tps",
            "type": "number"
          },
          "scale_mode": {
            "default": "linear",
            "enum": [
              "linear",
              "log"
            ],
            "title": "Scale Mode",
            "type": "string"
          },
          "scenario_id": {
            "description": "Scenario.id from /api/pipeline",
            "title": "Scenario Id",
            "type": "string"
          },
          "symbol": {
            "default": "DDOG",
            "description": "Ticker symbol (yfinance compatible)",
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9.\\-^=]+$",
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "default": "week",
            "enum": [
              "day",
              "week",
              "month"
            ],
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "scenario_id"
        ],
        "title": "AnalystStreamRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Elliott Wave Lab \u2014 pipeline + analyst HTTP API",
    "title": "EWL API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analyst/stream": {
      "post": {
        "description": "Server-Sent Events stream of an analyst narration.\n\nEvent types:\n    start      \u2014 { mode, model_id, scenario_id }\n    token      \u2014 { text } (whitespace-split, paced by rate_tps)\n    citations  \u2014 { citations[], cached, fell_back, model_id, prompt_version }\n    done       \u2014 { total_tokens, gen_ms } (gen_ms = real LLM wall-time, not playback)\n    error      \u2014 { message }",
        "operationId": "analyst_stream_api_v1_analyst_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalystStreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyst Stream",
        "tags": [
          "analyst"
        ]
      }
    },
...
  }
}